Add ability to publish programme to discord from ConClar
Read the program.json from a ConClar instance and publish it to Discord.
Assumes a forum per day, named after the day. Makes a forum thread for
each item. If the item is updated, it will edit the thread message. If
the change is deemed "significant" (time change, room change, or
cancelled), it is announced in a programme announcement channel.