Add tests to MembershipModule
This required splitting the module up into MembersApiClient and
MembershipChecker to make it easier to test.

I'm not particularly happy with the tests in MembershipChecker. I don't
think I've found the best way to test JDA yet. The intention is to avoid
hardcoding assumptions about how RestActions will be consumed (e.g.
queue vs complete) in order to make the tests less brittle. But I think
it's just ended up with an overly complicated fake.