Always check members on startup
Previously we were relying on the ready event to prompt us to check
members. This was always flaky, and because of some of the other code
changes the timings are slightly different, which means we aren't
getting the ready event reliably.

We now wait for ready before initializing the modules, so we can just
invoke the code when the module starts.