Send links to portal to verify users if missing portal record
Because the process of joining the discord through the portal is two
stage:
1) Get and store discord identity against membership
2) Send invite to discord

People can use different discord accounts for steps 1 and 2. To plug
this hole, we now generate a token in the portal that can be used to
associated a discord account with a member, and send a link to the
portal with that URL in a private thread to the user.