Replace .env with secrets.yaml file
We weren't really taking advantage of the .env files. Philosophically,
we've moved away from all secrets config via environment variables. We
were just using it as a way of loading the config from a file.

So replace it with a separate yaml file that we parse in the same way as
the config.