Make work with Plano style conclar JSON
Planorama uses `dateTime` instead of separate `date` and `time` fields,
and the tags are objects instead of strings.

Because we have a zoned date time coming in (as a string), this patch
swaps from using `LocalDateTime` to `ZonedDateTime` everywhere. When we
get to the database layer, it converts everything to a UTC string to
store in the database. This isn't the most correct strategy, as it can
lose information in certain cases. However, it isn't going to be a
problem in practice for regions with a straight forward offset with no
clock changes. We should improve it, but that would require putting more
complex logic into the queries.