Allow day channel names to be renamed
The day and day-tod resolver types now have a `dayMappings` property
that is a map of the English name to the channel name. This serves two
purposes:

1) It allows people to pick their own channel names. This would mainly
   be used for i18n
2) It lets us know what channels exist.

Before we assumed it could be any day of the week. This wasn't a
problem, because we only tried to resolve it if there was a programme
item scheduled on that day. However, now we are going to be adding
permission validations, we want to only check that the channels that
will be used exist.