Fixed everything being updated if timezone wasn't UTC
We were persisting the items using UTC in the database, and then
reading them back and comparing them to items with the configured
timezone.

Normalize everything in-memory to use the configured timezone, and only
store UTC in the database.

Note, we need to store it in UTC in the database because we do time
comparisons based on the strings, so we need a normalized form.