Validate channel permissions for programme module
Validate the channel permissions for the channels used in the programme
module on startup. This helps make sure the bot is configured correctly
to do all the things it needs to do.

Along with this, document more precisely what permissions are needed for
each channel in the config section of the README.